// apps/rf-terminal/src/app/features/quality/models/quality.models.ts
export type CriterionValue = 'SI' | 'NO' | 'NA';

export interface VerificationCriterion {
  id: string;
  label: string;
  sublabel?: string;
  value: CriterionValue;
  actionIfNo: string;
  responsible: string;
  instructionCode?: string;
  observations: string;
  isCritical?: boolean;
}